simplify python tests
